backend/requirements.txt
Django==5.2.6
Pillow==11.3.0
drf-spectacular==0.28.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.8.0
djoser==2.3.3
Faker==37.6.0
graphene==3.4.3
graphene-django==3.2.3
django-graphql-jwt==0.4.0
graphene-file-upload==1.3.0
django-storages==1.14.6
django-redis==6.0.0
gunicorn==21.2.0
celery
amqp
pika
psycopg2-binary